# web/src/api/openapi.merged.yaml
openapi: 3.1.0
info:
  title: ATRAG API
  description: ATRAG API Documentation
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /settings:
    get:
      summary: Get all settings
      description: Retrieve all application settings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A list of all settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
    put:
      summary: Update settings
      description: Update one or more application settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Settings'
      responses:
        '204':
          description: Settings updated successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /settings/test_mineru_token:
    post:
      summary: Test MinerU API Token
      description: Test the connectivity of the MinerU API Token
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Token test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  data:
                    type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
...
  /quotas:
    get:
      summary: Get user quotas
      description: Get quota information for the current user or specific user (admin only)
      tags:
        - quotas
      parameters:
        - name: user_id
          in: query
          description: User ID to get quotas for (admin only, defaults to current user)
          required: false
          schema:
            type: string
            example: user123
        - name: search
          in: query
          description: Search term for username, email, or user ID (admin only)
          required: false
          schema:
            type: string
            example: john
      responses:
        '200':
          description: Quota information retrieved successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/userQuotaInfo'
                  - $ref: '#/components/schemas/userQuotaList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '403':
          description: Forbidden - Admin access required for other users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /quotas/{user_id}:
    put:
      summary: Update user quota
      description: Update quota limit for a specific user (admin only)
      tags:
        - quotas
      parameters:
        - name: user_id
          in: path
          description: User ID to update quota for
          required: true
          schema:
            type: string
            example: user123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/quotaUpdateRequest'
      responses:
        '200':
          description: Quota updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/quotaUpdateResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /quotas/{user_id}/recalculate:
    post:
      summary: Recalculate user quota usage
      description: Recalculate and update current usage for all quota types for a user (admin only)
      tags:
        - quotas
      parameters:
        - name: user_id
          in: path
          description: User ID to recalculate quotas for
          required: true
          schema:
            type: string
            example: user123
      responses:
        '200':
          description: Quota usage recalculated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Quota usage recalculated successfully
                  updated_usage:
                    type: object
                    additionalProperties:
                      type: integer
                    example:
                      max_collection_count: 3
                      max_document_count: 25
                      max_bot_count: 2
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /system/default-quotas:
    get:
      summary: Get system default quotas
      description: Get system default quota configuration (admin only)
      tags:
        - quotas
      responses:
        '200':
          description: System default quotas retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/systemDefaultQuotasResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
    put:
      summary: Update system default quotas
      description: Update system default quota configuration (admin only)
      tags:
        - quotas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/systemDefaultQuotasUpdateRequest'
      responses:
        '200':
          description: System default quotas updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/systemDefaultQuotasUpdateResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
    CookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
  schemas:
...
    failResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: '400'
        message:
          type: string
          description: Error message
          example: Invalid request
...
    Settings:
      type: object
      properties:
        use_mineru:
          type: boolean
          description: Whether to use MinerU
        mineru_api_token:
          type: string
          description: API token for MinerU
        use_doc_ray:
          type: boolean
          description: Whether to use DocRay
        use_markitdown:
          type: boolean
          description: Whether to use MarkItDown
...
    quotaInfo:
      type: object
      description: Quota information for a specific quota type
      properties:
        quota_type:
          type: string
          description: Type of quota
          example: max_collection_count
        quota_limit:
          type: integer
          description: Maximum allowed usage
          example: 10
        current_usage:
          type: integer
          description: Current usage count
          example: 3
        remaining:
          type: integer
          description: Remaining quota available
          example: 7
      required:
        - quota_type
        - quota_limit
        - current_usage
        - remaining
    userQuotaInfo:
      type: object
      description: Complete quota information for a user
      properties:
        user_id:
          type: string
          description: User ID
          example: user123
        username:
          type: string
          description: Username
          example: john_doe
        email:
          type: string
          description: User email
          example: john@example.com
          nullable: true
        role:
          type: string
          description: User role
          example: rw
        quotas:
          type: array
          description: List of quota information
          items:
            $ref: '#/components/schemas/quotaInfo'
      required:
        - user_id
        - role
        - quotas
    userQuotaList:
      type: object
      description: List of user quota information (admin view)
      properties:
        items:
          type: array
          description: List of user quota information
          items:
            $ref: '#/components/schemas/userQuotaInfo'
      required:
        - items
    quotaUpdateRequest:
      type: object
      description: Request to update user quotas (supports both single and batch updates)
      properties:
        max_collection_count:
          type: integer
          description: New limit for collection count
          minimum: 0
          nullable: true
        max_document_count:
          type: integer
          description: New limit for document count
          minimum: 0
          nullable: true
        max_document_count_per_collection:
          type: integer
          description: New limit for documents per collection
          minimum: 0
          nullable: true
        max_bot_count:
          type: integer
          description: New limit for bot count
          minimum: 0
          nullable: true
      example:
        max_collection_count: 20
        max_document_count: 2000
        max_bot_count: 10
    quotaUpdateResponse:
      type: object
      description: Response after updating user quotas (supports both single and batch updates)
      properties:
        success:
          type: boolean
          description: Whether the update was successful
          example: true
        message:
          type: string
          description: Status message
          example: Quotas updated successfully
        user_id:
          type: string
          description: User ID that was updated
          example: user123
        updated_quotas:
          type: array
          description: List of updated quotas
          items:
            type: object
            properties:
              quota_type:
                type: string
                description: Type of quota that was updated
                example: max_collection_count
              old_limit:
                type: integer
                description: Previous quota limit
                example: 10
              new_limit:
                type: integer
                description: New quota limit
                example: 20
            required:
              - quota_type
              - old_limit
              - new_limit
      required:
        - success
        - message
        - user_id
        - updated_quotas
    systemDefaultQuotas:
      type: object
      description: System default quota configuration
      properties:
        max_collection_count:
          type: integer
          description: Default maximum collection count
          example: 10
          minimum: 0
        max_document_count:
          type: integer
          description: Default maximum document count
          example: 1000
          minimum: 0
        max_document_count_per_collection:
          type: integer
          description: Default maximum documents per collection
          example: 100
          minimum: 0
        max_bot_count:
          type: integer
          description: Default maximum bot count
          example: 5
          minimum: 0
      required:
        - max_collection_count
        - max_document_count
        - max_document_count_per_collection
        - max_bot_count
    systemDefaultQuotasResponse:
      type: object
      description: Response containing system default quotas
      properties:
        quotas:
          $ref: '#/components/schemas/systemDefaultQuotas'
      required:
        - quotas
    systemDefaultQuotasUpdateRequest:
      type: object
      description: Request to update system default quotas
      properties:
        quotas:
          $ref: '#/components/schemas/systemDefaultQuotas'
      required:
        - quotas
    systemDefaultQuotasUpdateResponse:
      type: object
      description: Response after updating system default quotas
      properties:
        success:
          type: boolean
          description: Whether the update was successful
          example: true
        message:
          type: string
          description: Status message
          example: System default quotas updated successfully
        quotas:
          $ref: '#/components/schemas/systemDefaultQuotas'
      required:
        - success
        - message
        - quotas
...